+++ ucc/__init__.py
@@ -1,4 +1,3 @@
 from .transpilers import UCCTranspiler
 from .compile import compile
-from .custom_cx import CXCancellation
 from .quantum_translator import QuantumTranslator
