+++ ucc/transpiler_passes/__init__.py
@@ -7,3 +7,4 @@ from .unitary_synthesis import UnitarySynthesis
 from .collect_2q_blocks import Collect2qBlocks
 from .collect_1q_runs import Collect1qRuns
 from .optimize_1q_decomposition import Optimize1qGatesDecomposition
+from .custom_cx import CXCancellation
